Refactors GCP error handling for improved clarity and context
Reorganizes GCP error handling by introducing a new `errors.go` file,
replacing verbose error messages with a concise `briefGcpError` type
that prioritizes displaying relevant error messages. Enhances error
annotations with additional context for forbidden access errors,
directing users to verify project ID and credentials. Removes
unused project name extraction logic and the deprecated `utils.go` file.

Improves test coverage with updates to the renamed test file, ensuring
proper error wrapping and unwrapping behavior.

Relates to issue #1519.
